+++ src/main/java/domain/Car.java
@@ -4,13 +4,12 @@ import utils.NumberGenerator;
 
 public class Car {
 
-    private final String name;
+    private final CarName name;
     private int drivenDistance = 0;
     private final NumberGenerator numberGenerator;
 
     public Car(final String name, NumberGenerator numberGenerator) {
-        validateLength(name);
-        this.name = name;
+        this.name = new CarName(name);
         this.numberGenerator = numberGenerator;
     }
 
@@ -33,17 +32,6 @@ public class Car {
     }
 
     public String getName() {
-        return name;
-    }
-
-    private void validateLength(final String carName) {
-        final String NOT_PROPER_CAR_NAME_LENGTH = "[ERROR] 자동차 이름 길이는 %d자 이상, %d자 이하여야합니다.";
-        final int MIN_CAR_NAME_LENGTH = 1;
-        final int MAX_CAR_NAME_LENGTH = 5;
-
-        if (carName.length() < MIN_CAR_NAME_LENGTH || carName.length() > MAX_CAR_NAME_LENGTH) {
-            throw new IllegalArgumentException(
-                String.format(NOT_PROPER_CAR_NAME_LENGTH, MIN_CAR_NAME_LENGTH, MAX_CAR_NAME_LENGTH));
-        }
+        return name.getName();
     }
 }
