+++ src/main/java/domain/CarName.java
@@ -1,0 +1,26 @@
+package domain;
+
+public class CarName {
+
+    private final String name;
+
+    public CarName(final String name) {
+        validateLength(name);
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    private void validateLength(final String carName) {
+        final String NOT_PROPER_CAR_NAME_LENGTH = "[ERROR] 자동차 이름 길이는 %d자 이상, %d자 이하여야합니다.";
+        final int MIN_CAR_NAME_LENGTH = 1;
+        final int MAX_CAR_NAME_LENGTH = 5;
+
+        if (carName.length() < MIN_CAR_NAME_LENGTH || carName.length() > MAX_CAR_NAME_LENGTH) {
+            throw new IllegalArgumentException(
+                String.format(NOT_PROPER_CAR_NAME_LENGTH, MIN_CAR_NAME_LENGTH, MAX_CAR_NAME_LENGTH));
+        }
+    }
+}
